// src/providers/types/RedPacket.ts
export enum PlatformType {
    lens = 'lens',
    farcaster = 'farcaster',
    twitter = 'twitter',
}

export enum ActionType {
    Send = 'send',
    Claim = 'claim',
}

export enum SourceType {
    All = 'all',
    FireflyAPP = 'firefly_app',
    FireflyPC = 'firefly_pc',
    MaskNetwork = 'mask_network',
}

export enum RedPacketStatus {
    View = 'VIEW',
    Refunding = 'REFUNDING',
    Refund = 'REFUND',
    Expired = 'EXPIRED',
    Empty = 'EMPTY',
    Send = 'SEND',
}

export enum StrategyType {
    profileFollow = 'profileFollow',
    postReaction = 'postReaction',
    nftOwned = 'nftOwned',
}

export interface StrategyPayload {
    type: StrategyType;
    payload: Array<ProfileFollowStrategyPayload | NftOwnedStrategyPayload> | PostReactionStrategyPayload;
}

export interface ProfileFollowStrategyPayload {
    platform: PlatformType;
    profileId: string;
}

export interface PostReactionStrategyPayload {
    params?: Array<{
        platform: PlatformType;
        postId: string;
        handle?: string;
    }>;
    reactions: string[];
}

export interface NftOwnedStrategyPayload {
    chainId: number;
    contractAddress: string;
}